+++ web/index.service.worker.js
@@ -4,7 +4,7 @@
 // Incrementing CACHE_VERSION will kick off the install event and force
 // previously cached resources to be updated from the network.
 /** @type {string} */
-const CACHE_VERSION = '1786733003|8583062';
+const CACHE_VERSION = '1786738738|6277987';
 /** @type {string} */
 const CACHE_PREFIX = 'No Hitter-sw-cache-';
 const CACHE_NAME = CACHE_PREFIX + CACHE_VERSION;